requirements.txt
fastapi
psycopg2-binary 
PyJWT
PyJWT[crypto]
python-dotenv
tenacity
requests
uvicorn[standard]
